+++ src/gui/qt/main_window.cpp
@@ -52,7 +52,7 @@
 #include "gui/query/types/change_game_query.h"
 #include "gui/query/types/clear_all_metadata_query.h"
 #include "gui/query/types/clear_plugin_metadata_query.h"
-#include "gui/query/types/get_conflicting_plugins_query.h"
+#include "gui/query/types/get_overlapping_plugins_query.h"
 #include "gui/query/types/get_game_data_query.h"
 #include "gui/query/types/sort_plugins_query.h"
 #include "gui/version.h"
@@ -1027,9 +1027,9 @@ void MainWindow::setFiltersState(PluginFiltersState&& filtersState) {
 
 void MainWindow::setFiltersState(
     PluginFiltersState&& filtersState,
-    std::vector<std::string>&& conflictingPluginNames) {
+    std::vector<std::string>&& overlappingPluginNames) {
   proxyModel->setFiltersState(std::move(filtersState),
-                              std::move(conflictingPluginNames));
+                              std::move(overlappingPluginNames));
 
   updateCounts(pluginItemModel->getGeneralMessages(),
                pluginItemModel->getPluginItems());
@@ -1160,7 +1160,7 @@ void MainWindow::showFirstRunDialog() {
   auto listItem1 =
       boost::locale::translate(
           "CRCs are only displayed after plugins have been loaded, either by "
-          "conflict filtering, or by sorting.")
+          "overlap filtering, or by sorting.")
           .str();
 
   auto listItem2 =
@@ -1396,7 +1396,7 @@ bool MainWindow::handlePluginsSorted(std::vector<QueryResult> results) {
     handleMasterlistUpdated(results);
   }
 
-  filtersWidget->resetConflictsAndGroupsFilters();
+  filtersWidget->resetOverlapAndGroupsFilters();
 
   auto sortedPlugins = std::get<PluginItems>(results.back());
 
@@ -2338,7 +2338,7 @@ void MainWindow::on_filtersWidget_pluginFilterChanged(
   setFiltersState(std::move(filtersState));
 }
 
-void MainWindow::on_filtersWidget_conflictsFilterChanged(
+void MainWindow::on_filtersWidget_overlapFilterChanged(
     std::optional<std::string> targetPluginName) {
   try {
     if (!targetPluginName.has_value()) {
@@ -2350,15 +2350,15 @@ void MainWindow::on_filtersWidget_conflictsFilterChanged(
       return;
     }
 
-    handleProgressUpdate(translate("Identifying conflicting plugins..."));
+    handleProgressUpdate(translate("Identifying overlapping plugins..."));
 
-    std::unique_ptr<Query> query = std::make_unique<GetConflictingPluginsQuery>(
+    std::unique_ptr<Query> query = std::make_unique<GetOverlappingPluginsQuery>(
         state.GetCurrentGame(),
         state.getSettings().getLanguage(),
         targetPluginName.value());
 
     executeBackgroundQuery(
-        std::move(query), &MainWindow::handleConflictsChecked, nullptr);
+        std::move(query), &MainWindow::handleOverlapFilterChecked, nullptr);
   } catch (const std::exception& e) {
     handleException(e);
   }
@@ -2461,7 +2461,7 @@ void MainWindow::on_searchDialog_currentResultChanged(size_t resultIndex) {
 void MainWindow::handleGameChanged(QueryResult result) {
   try {
     filtersWidget->setGameId(state.GetCurrentGame().GetSettings().Id());
-    filtersWidget->resetConflictsAndGroupsFilters();
+    filtersWidget->resetOverlapAndGroupsFilters();
     disablePluginActions();
 
     handleGameDataLoaded(result);
@@ -2666,21 +2666,21 @@ void MainWindow::handleMasterlistsUpdated(std::vector<QueryResult> results) {
   }
 }
 
-void MainWindow::handleConflictsChecked(QueryResult result) {
+void MainWindow::handleOverlapFilterChecked(QueryResult result) {
   try {
     progressDialog->reset();
 
     // The result is not an array of PluginItem objects.
     // Instead it's an array of objects, with each having a metadata property
-    // that is a PluginItem object, and a conflicts property that is
+    // that is a PluginItem object, and an overlap property that is
     // a boolean.
     std::vector<PluginItem> gameDataLoadedResult;
-    std::vector<std::string> conflictingPluginNames;
+    std::vector<std::string> overlappingPluginNames;
 
     for (const auto& pluginPair :
-         std::get<GetConflictingPluginsResult>(result)) {
+         std::get<GetOverlappingPluginsResult>(result)) {
       if (pluginPair.second) {
-        conflictingPluginNames.push_back(pluginPair.first.name);
+        overlappingPluginNames.push_back(pluginPair.first.name);
       }
 
       gameDataLoadedResult.push_back(pluginPair.first);
@@ -2689,7 +2689,7 @@ void MainWindow::handleConflictsChecked(QueryResult result) {
     handleGameDataLoaded(gameDataLoadedResult);
 
     setFiltersState(filtersWidget->getPluginFiltersState(),
-                    std::move(conflictingPluginNames));
+                    std::move(overlappingPluginNames));
 
     // Load order state was refreshed when plugins were loaded, so check for
     // ambiguity.