+++ docs/app/changelog.rst
@@ -19,6 +19,9 @@ Changed
   previously referred to as the "override" plugin type and flag is now
   referred to as the "overlay" plugin type and flag for consistency with xEdit,
   Wrye Bash and Mod Organizer 2.
+- The "Show only conflicting plugins for" filter has been renamed to
+  "Show only overlapping plugins for" for clarity and consistency with other
+  LOOT functionality.
 
 0.22.0 - 2023-10-07
 ===================